fix(): Setup DNS config in resolv.conf (#2)
Backed up the original resolv.conf and overwrote its contents to
redirect dns queries to the cmd-nsc sidecar.
This was needed because istio-proxy that runs as one of the
sidecars seems to read the resolv.conf during bootup and cache the config.
Any subsequent change to the resolv.conf is not picked up by the proxy. So
if the cmd-nsc sidecar takes time to boot up and modify the resolv.conf for
all the other containers, the istio-proxy sidecar would have read the
original resolv.conf and ignored the config written by cmd-nsc. This causes
name resolution to fail in istio-proxy for the domains that are serviced by
a custom dns server other than the default kube-dns.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>