fix(): Apply workaround to nsmgr tcp connection
The nsmgr daemonset is exposed through a k8s service with
internalTrafficPolicy set to "Local" so that clients can connect only
to the node local nsmgr endpoint pod. This feature is not working
reliably as the connection is sometimes made to a remote node endpoint.
To workaround this problem, a service is created for every node in the
cluster and the cmd-nsc is made to connect to its node local service.

This fix should be removed once the internalTrafficPolicy feature is GA
and works reliably.

Also, removed the code to overwrite the resolv.conf. It was added to
handle a specific version mismatch between istiod and istio-proxy
sidecar. We no longer support mismatch in versions, we expect both to be
on the same version.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>