fix(): Modify resolv.conf to workaround istio dns issue
Istio-proxy when run in router mode caches the dns server IP in the
resolv.conf when it boots up, and since the resolv.conf contains the
IP of kube-dns, istio-proxy fails to resolve any non cluster.local
names. The cmd-nsc-init would overwrite the resolv.conf and add the
localhost IP as one of the nameservers before istio-proxy boots up so
that it caches the localhost IP and resolves non cluster.local names
through the cmd-nsc nsm sidecar dns proxy.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>